Fix all compiler warnings from various build systems.
Fix vsnprintf() implementation to properly handle long and long long ints.


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.1@5845 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ test/scroll.cxx
@@ -71,7 +71,7 @@ void box_cb(Fl_Widget* o, void*) {
 }
 
 void type_cb(Fl_Widget*, void* v) {
-  thescroll->type((long)v);
+  thescroll->type((uchar)((long)v));
   thescroll->redraw();
 }
 
@@ -87,7 +87,7 @@ Fl_Menu_Item choices[] = {
 };
 
 void align_cb(Fl_Widget*, void* v) {
-  thescroll->scrollbar.align((long)v);
+  thescroll->scrollbar.align((uchar)((long)v));
   thescroll->redraw();
 }
